backend/requirements.txt
Flask
Flask-CORS
flask-session
requests
config
python-dotenv